fix: Update warning path
